Compiler API improvements:
Added a new FileSpec class whose initializer requires a file path or the source code (and can take both). Updated the parser to read files from disk only when the source code is not passed. The upshot is that files to be compiled can be passed in memory, with no disk I/O requirement.

Renamed CobraParser.parseSource to .parse.

Renamed Compiler.compileFilesNamed to .compile.

Added a new CompileParams class that simplifies the invocation of .compile. This also includes a new ability to explicitly specify the phases.